+++ 2021/15_2-pass_gamut_boundary/jzazbz_azbz_czhz_plot.py
@@ -39,104 +39,6 @@ __email__ = 'toru.ver.11 at-sign gmail.com'
 __all__ = []
 
 
-def calc_chroma_boundary_specific_ligheness_jzazbz_method_c(
-        lch, cs_name, peak_luminance, c0):
-    """
-    parameters
-    ----------
-    lightness : float
-        lightness value(Jzazbz). range is 0.0 - 1.0.
-    hue_sample : int
-        Sample number of the Hue
-    cs_name : string
-        A color space name. ex. "ITU-R BT.709", "ITU-R BT.2020"
-
-    Examples
-    --------
-    >>> boundary_jch = calc_chroma_boundary_specific_ligheness_jzazbz(
-    ...     lightness=0.5, hue_sample=16, cs_name=cs.BT2020,
-    ...     peak_luminance=10000)
-    [[  5.00000000e-01   2.72627831e-01   0.00000000e+00]
-     [  5.00000000e-01   2.96944618e-01   2.40000000e+01]
-     [  5.00000000e-01   3.19167137e-01   4.80000000e+01]
-     [  5.00000000e-01   2.51322746e-01   7.20000000e+01]
-     [  5.00000000e-01   2.41002083e-01   9.60000000e+01]
-     [  5.00000000e-01   2.76854515e-01   1.20000000e+02]
-     [  5.00000000e-01   3.99024010e-01   1.44000000e+02]
-     [  5.00000000e-01   2.64456749e-01   1.68000000e+02]
-     [  5.00000000e-01   2.32390404e-01   1.92000000e+02]
-     [  5.00000000e-01   2.51740456e-01   2.16000000e+02]
-     [  5.00000000e-01   3.38995934e-01   2.40000000e+02]
-     [  5.00000000e-01   3.09918404e-01   2.64000000e+02]
-     [  5.00000000e-01   2.71250725e-01   2.88000000e+02]
-     [  5.00000000e-01   2.59991646e-01   3.12000000e+02]
-     [  5.00000000e-01   2.63157845e-01   3.36000000e+02]
-     [  5.00000000e-01   2.72627831e-01   3.60000000e+02]]
-    """
-    # lch --> rgb
-
-    jj = lch[..., 0]
-    chroma_init = lch[..., 1]
-    hue = np.deg2rad(lch[..., 2])
-
-    trial_num = 20
-
-    r_val = chroma_init
-
-    for t_idx in range(trial_num):
-        aa = r_val * np.cos(hue)
-        bb = r_val * np.sin(hue)
-        jzazbz = tstack((jj, aa, bb))
-        large_xyz = jzazbz_to_large_xyz(jzazbz)
-        rgb_luminance = XYZ_to_RGB(
-            large_xyz, cs.D65, cs.D65,
-            RGB_COLOURSPACES[cs_name].matrix_XYZ_to_RGB)
-
-        ng_idx = is_out_of_gamut_rgb(rgb=rgb_luminance/peak_luminance)
-        ok_idx = np.logical_not(ng_idx)
-        add_sub = c0 / (2 ** (t_idx))
-        r_val[ok_idx] = r_val[ok_idx] + add_sub
-        r_val[~ok_idx] = r_val[~ok_idx] - add_sub
-
-    zero_idx = (chroma_init <= 0)
-    r_val[zero_idx] = 0.0
-
-    jzczhz = tstack([jj, r_val, np.rad2deg(hue)])
-
-    return jzczhz
-
-
-def create_jzazbz_gamut_boundary_method_c(
-        hue_sample=8, lightness_sample=8, chroma_sample=1024,
-        color_space_name=cs.BT709, luminance=100):
-
-    c0 = JZAZBZ_CHROMA_MAX / (chroma_sample - 1)
-    # create 2d lut using method B
-    create_jzazbz_gamut_boundary_lut_type2(
-        hue_sample=hue_sample, lightness_sample=lightness_sample,
-        chroma_sample=chroma_sample, color_space_name=color_space_name,
-        luminance=luminance)
-
-    lut_b = np.load(
-        make_jzazbz_gb_lut_fname_methodb_b(
-            color_space_name=color_space_name, luminance=luminance,
-            lightness_num=lightness_sample, hue_num=hue_sample))
-
-    # create 2d lut using method C
-    lut_c = np.zeros_like(lut_b)
-    for l_idx in range(lightness_sample):
-        jzczhz_init = lut_b[l_idx]
-        jzczhz = calc_chroma_boundary_specific_ligheness_jzazbz_method_c(
-            lch=jzczhz_init, cs_name=color_space_name,
-            peak_luminance=luminance, c0=c0)
-        lut_c[l_idx] = jzczhz
-
-    fname = make_jzazbz_gb_lut_fname_method_c(
-        color_space_name=color_space_name, luminance=luminance,
-        lightness_num=lightness_sample, hue_num=hue_sample)
-    np.save(fname, np.float32(lut_c))
-
-
 def debug_plot_azbz_plane_with_interpolation(
         hue_sample=256, lightness_sample=256, j_num_intp=256,
         color_space_name=cs.BT2020, luminance=10000):
@@ -353,18 +255,10 @@ def plot_cj_plane_with_interpolation(
             pool.map(thread_wrapper_plot_cj_plane_with_interpolation, args)
 
 
-def debug_plot_jzazbz():
-    chroma_sample = 4
-    hue_sample = 256
-    lightness_sample = 128
-    luminance = 100
-    h_num_intp = 100
-    j_num_intp = 100
-    color_space_name = cs.BT709
-    create_jzazbz_gamut_boundary_method_c(
-        hue_sample=hue_sample, lightness_sample=lightness_sample,
-        chroma_sample=chroma_sample,
-        color_space_name=color_space_name, luminance=luminance)
+def debug_plot_jzazbz(
+        hue_sample=256, lightness_sample=128,
+        luminance=100, h_num_intp=100, j_num_intp=100,
+        color_space_name=cs.BT709):
     debug_plot_azbz_plane_with_interpolation(
         hue_sample=hue_sample, lightness_sample=lightness_sample,
         j_num_intp=j_num_intp,
